Rename `java.scala` -> `JavaConverters.scala`
